+++ changelog.md
@@ -1,5 +1,10 @@
 # PSBackup Change Log
 
+## February 8. 2020
+
++ Bug fixes in `WeeklyBackup.ps1` from changes I made to variable names and paths
++ Moved `RarBackup.ps1` to this repository
+
 ## February 7, 2020
 
 + Initial file upload